+++ CHANGES.md
@@ -7,6 +7,7 @@
 which included change [CLJS-1367](https://github.com/clojure/clojurescript/commit/9e662c3ea5b9f536f303e9084415e988bf5d0878), should work.
 - Expose ClojureScript compiler options in `.cljs.edn` Boot fileset metadata for
 others tasks (e.g. boot-reload)
+- Hide stacktraces for Cljs compiler exceptions (reader exception, analysis error)
 
 ## 1.7.228-2 (18.10.2016)